big time refactoring but it's all good now
