+++ internal/algo/algo.go
